backend/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
python-multipart==0.0.12
pillow==11.0.0
